IReadOnlyDictionaryEqualByComparer
